+++ src/IR/IR.ts
@@ -35,6 +35,7 @@ import {
 } from "./loops";
 import { Argv, Identifier, IntegerLiteral, StringLiteral } from "./terminals";
 import { Block, IfStatement, VarDeclaration, Variants } from "./toplevel";
+import { simpleType, ValueType } from "./types";
 
 export * from "./assignments";
 export * from "./collections";
@@ -42,6 +43,7 @@ export * from "./exprs";
 export * from "./loops";
 export * from "./terminals";
 export * from "./toplevel";
+export * from "./types";
 
 export interface BaseExpr {
   valueType?: ValueType;
@@ -86,139 +88,6 @@ export type Expr =
   | ConditionalOp
   | ManyToManyAssignment
   | OneToManyAssignment;
-
-/** The type of the value of a node when evaluated */
-export type ValueType =
-  | { type: "void" }
-  | { type: "number" }
-  | { type: "string" }
-  | { type: "boolean" }
-  | { type: "List"; member: ValueType }
-  | { type: "Table"; key: "number" | "string"; value: ValueType }
-  | { type: "Array"; member: ValueType; length: number }
-  | { type: "Set"; member: ValueType };
-
-export function simpleType(
-  type: "void" | "number" | "string" | "boolean"
-): ValueType {
-  return { type };
-}
-export function tableType(
-  key: "number" | "string",
-  value: ValueType | "void" | "number" | "string" | "boolean"
-): ValueType {
-  return {
-    type: "Table",
-    key,
-    value: typeof value === "string" ? simpleType(value) : value,
-  };
-}
-export function setType(
-  member: ValueType | "void" | "number" | "string" | "boolean"
-): ValueType {
-  return {
-    type: "Set",
-    member: typeof member === "string" ? simpleType(member) : member,
-  };
-}
-export function listType(
-  member: ValueType | "void" | "number" | "string" | "boolean"
-): ValueType {
-  return {
-    type: "List",
-    member: typeof member === "string" ? simpleType(member) : member,
-  };
-}
-export function arrayType(
-  member: ValueType | "void" | "number" | "string" | "boolean",
-  length: number
-): ValueType {
-  return {
-    type: "Array",
-    member: typeof member === "string" ? simpleType(member) : member,
-    length,
-  };
-}
-
-function getType(expr: Expr, program: Program): ValueType {
-  if (expr.valueType === null) {
-    expr.valueType = calcType(expr, program);
-  }
-  return expr.valueType!;
-}
-
-export function calcType(expr: Expr, program: Program): ValueType {
-  switch (expr.type) {
-    case "Argv":
-      return listType("string");
-    case "VarDeclaration":
-      return simpleType("void");
-    case "Assignment":
-      return getType(expr.expr, program);
-    case "Print":
-      return simpleType("void");
-    case "FunctionCall":
-      return simpleType("void"); // TODO
-    case "MethodCall":
-      return simpleType("void"); // TODO
-    case "BinaryOp":
-      return simpleType("number"); // TODO
-    case "UnaryOp":
-      return simpleType("number"); // TODO
-    case "Identifier":
-      if (program.variables.has(expr.name)) {
-        return program.variables.get(expr.name)!;
-      }
-      throw new Error(`Undeclared variable ${expr.name} encountered!`);
-    case "StringLiteral":
-      return simpleType("string");
-    case "IntegerLiteral":
-      return simpleType("number");
-    case "ArrayConstructor":
-      return arrayType(getType(expr.exprs[0], program), expr.exprs.length);
-    case "ListConstructor":
-      return listType(getType(expr.exprs[0], program));
-    case "StringGetByte":
-      return simpleType("string");
-    case "TableGet": {
-      const tableType = getType(expr.table, program);
-      if (tableType.type !== "Table") {
-        throw new Error("Type of TableGet must be used on a table");
-      }
-      return tableType.value;
-    }
-    case "TableSet":
-      return simpleType("void");
-    case "ArrayGet": {
-      const arrType = getType(expr.array, program);
-      if (arrType.type !== "Array") {
-        throw new Error("Type of TableGet must be used on a table");
-      }
-      return arrType.member;
-    }
-    case "ArraySet":
-      return simpleType("void");
-    case "ListGet": {
-      const listType = getType(expr.list, program);
-      if (listType.type !== "List") {
-        throw new Error("Type of TableGet must be used on a table");
-      }
-      return listType.member;
-    }
-    case "ListSet":
-      return simpleType("void");
-    case "ListPush":
-      return simpleType("void");
-    case "MutatingBinaryOp":
-      return simpleType("void");
-    case "ConditionalOp":
-      return getType(expr.consequent, program);
-    case "ManyToManyAssignment":
-      return simpleType("void");
-    case "OneToManyAssignment":
-      return getType(expr.expr, program);
-  }
-}
 
 /**
  * Program node. This should be the root node. Raw OK
